envoie bd

+++ Projet_interface_BD/BD_ProjetBlazor/BD_ProjetBlazor/Services/Requete_EntreeSortieService.cs
@@ -58,7 +58,7 @@ namespace BD_ProjetBlazor.Services
 
             if (stationnement != null)
             {
-                return stationnement.Tarif;  
+                return (decimal)stationnement.Tarif;  
             }
             else
             {
